Refactor modal component
