+++ src/components/Global/Modal/Modal.module.scss
@@ -1,0 +1,63 @@
+@use "@/styles/constants";
+
+.overlay {
+    --animation-time: 0.4s;
+    --opacity: 0.5;
+
+    display: flex;
+    position: fixed;
+    top: 0;
+    justify-content: center;
+    align-items: center;
+    z-index: constants.$modal-z;
+    animation-duration: var(--animation-time);
+    animation-name: fadein;
+    background-color: rgba(0, 0, 0, var(--opacity));
+    width: 100lvw;
+    height: 100lvh;
+
+    .modal {
+        --border-rad: 20px;
+        --height: 85%;
+        --padding: 1rem;
+        --width: 90%;
+
+        border-radius: var(--border-rad);
+        background: var(--theme-accent);
+        padding: var(--padding);
+        width: var(--width);
+        height: var(--height);
+        overflow-y: scroll;
+    }
+}
+
+.closing {
+    animation-duration: var(--animation-time);
+    animation-name: fadeout;
+}
+
+@keyframes fadein {
+    from {
+        opacity: 0;
+    }
+
+    to {
+        opacity: 1;
+    }
+}
+
+@keyframes fadeout {
+    from {
+        opacity: 1;
+    }
+
+    to {
+        opacity: 0;
+    }
+}
+
+@media screen and (min-width: constants.$desktop) {
+    .overlay {
+        display: none;
+    }
+}
